Restaurant Bot/requirements.txt
python-telegram-bot==13.13
pandas==1.5.0
numpy==1.23.4
